Commented changes on resources, because the compiler is complaining about consts. Apparently, it is a C89 limitation, although I've been able to use this same condition in the same file in the code... I will be changing the array into a 'movable' linked list.

+++ os/platform/pic18f25k80/system_counter.c
@@ -19,5 +19,5 @@ void PlatformInitializeSystemCounter() {
 }
 
 uint8_t PlatformHasInterruptSystemCounter() {
-	return INTCONbits.TMR0IE && INTCONbits.TMR0IF;
+	return (uint8_t)(INTCONbits.TMR0IE && INTCONbits.TMR0IF);
 }
